compilation fixes
