+++ Searchable.h
@@ -7,7 +7,10 @@
 
 
 #include <vector>
+#include <memory>
 #include "State.h"
+#include "Util.h"
+
 
 template <typename S>
 class Searchable {
@@ -21,9 +24,8 @@ public:
     virtual std::vector<State<S>*>* getAllPossibleStates() = 0;
     virtual void setAllPossibleStates() = 0;
     virtual void updateState(State<S> *next) = 0;
-    virtual bool visited(S* state) = 0;
+    virtual bool visited(State<S>* state) = 0;
     virtual void setVisit(State<S>* state) = 0;
-    virtual void setFatherstate() = 0;
     virtual State<S>* getState() = 0;
 };
